requirements.txt
Django==1.10
docker-compose==1.8
GitPython==2.1.3
